server/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
asyncio==3.4.3
bidict==0.22.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
dnspython==1.16.0
eventlet==0.33.0
Flask==2.1.1
Flask-SocketIO==5.1.1
greenlet==1.1.2
grpcio==1.44.0
grpcio-tools==1.44.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
protobuf==3.20.0
pyparsing==3.0.8
python-engineio==4.3.1
python-socketio==5.5.2
redis==4.2.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
vine==5.0.0
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.1
wrapt==1.14.0
